Move Amino message types and converters to module folders
