+++ packages/stargate/src/modules/authz/aminomessages.ts
@@ -1,0 +1,9 @@
+import { AminoConverters } from "../../aminoconverters";
+
+export function createAuthzAminoConverters(): AminoConverters {
+  return {
+    "/cosmos.authz.v1beta1.MsgGrant": "not_supported_by_chain",
+    "/cosmos.authz.v1beta1.MsgExec": "not_supported_by_chain",
+    "/cosmos.authz.v1beta1.MsgRevoke": "not_supported_by_chain",
+  };
+}
