+++ packages/stargate/src/modules/feegrant/aminomessages.ts
@@ -1,0 +1,8 @@
+import { AminoConverters } from "../../aminoconverters";
+
+export function createFreegrantAminoConverters(): AminoConverters {
+  return {
+    "/cosmos.feegrant.v1beta1.MsgGrantAllowance": "not_supported_by_chain",
+    "/cosmos.feegrant.v1beta1.MsgRevokeAllowance": "not_supported_by_chain",
+  };
+}
